fix: address CI lint and security findings on ReconnectDialer
- gocyclo: extract dial-config parsing from LoadConfig into a
  helper parseDialConfig() to drop cyclomatic complexity below 15.
- unconvert: drop redundant int32(n) cast in recordingDialer
  test helper (n is already int32 from atomic.Int32.Add).
- unused: remove dead rngOnce/rng fields from ReconnectDialer.
  The struct was scaffolded with thread-safe RNG state but the
  final implementation uses math/rand/v2 package-level functions
  (which are concurrent-safe in v2), so the fields were never
  read or written.
- gosec G404: switch the inline directive from //nolint:gosec
  to // #nosec G404 — the standalone gosec runner does not honor
  golangci-lint nolint comments; #nosec is the native form.
  Jitter is not security-sensitive (it perturbs backoff timing
  to avoid retry stampedes), so weak RNG is acceptable here.

All four findings are local to this branch; no existing code touched.

+++ internal/proxy/reconnect_test.go
@@ -24,7 +24,7 @@ type recordingDialer struct {
 
 func (r *recordingDialer) Dial(_ context.Context, _, _ string) (net.Conn, error) {
 	n := r.attempts.Add(1)
-	if int32(n) <= r.failuresBefore {
+	if n <= r.failuresBefore {
 		return nil, r.errFail
 	}
 	a, b := net.Pipe()
